Dynamically crate scores
