+++ src/scoreboard/fetch_game.js
@@ -3,10 +3,10 @@ const fetchGame = async () => {
     'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/XcndOEVAK91fzDURy1fY/scores',
   ).then((response) => response.json());
 
-  const length = result.result.length
+  const length = (result.result.length - 1)
 
-  for(let i = 0; i < (length - 1); i += 1){
-    for(let j = i; j < (length - 1); j += 1){
+  for(let i = 0; i < length; i += 1){
+    for(let j = 0; j < (length - i); j += 1){
       if(result.result[j].score < result.result[j + 1].score){
         let my_var = result.result[j];
         result.result[j] = result.result[j + 1];
