+++ src/scoreboard/gameover.js
@@ -1,42 +1,16 @@
-import gameWorld from '../world/config';
-import fetchGame from './fetch_game';
+import gameOverScreen from './go_screen';
+import createScore from './create_score';
 
-const GameOver = async () => {
+const GameOver = () => {
 
-  let myArr = await fetchGame();
+  const name = localStorage.getItem('player');
+  const score = localStorage.getItem('score');
+
+  createScore(name, score);
+
+  gameOverScreen();
 
   //global.game.destroy(true, false);
-
-  const display = document.createElement('div');
-  display.id = 'display';
-
-  const p = document.createElement('p');
-  p.classList.add
-  p.innerHTML = 'Game Over!';
-
-  display.appendChild(p);
-
-  const scoreBoard = document.createElement('ul');
-
-  for(let i = 0; i < 5; i += 1){
-    const scoreItem = document.createElement('li');
-    scoreItem.innerHTML = 'Player: ' + myArr[i].user + ', score: ' + myArr[i].score;
-    scoreBoard.appendChild(scoreItem);
-  }
-
-  display.appendChild(scoreBoard);
-
-  const button = document.createElement('div');
-  button.innerHTML = 'Play again';
-  button.id = 'gameStart';
-
-  button.addEventListener('click', () => {
-    gameWorld();
-  });
-
-  display.appendChild(button);
-
-  document.body.appendChild(display);
 
 }
